Moving add modal to a new file
